requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
datetime
pillow>=10.0.0
streamlit-option-menu>=0.3.6
requests>=2.31.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
supabase>=2.3.0
python-dotenv>=1.0.0 